- bugfix: change nfs-server write method: async -> sync (at /etc/exports) - simplify shell scripts.

+++ src/update-configMap-cinder-setup.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-kubectl delete configmap cinder-setup
+if (kubectl get configmap | grep -q "cinder-setup"); then
+	kubectl delete configmap cinder-setup
+fi
 kubectl create -f configMap-cinder-setup.yaml
