+++ src/update-configMap-env-common.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-kubectl delete configmap env-common
+if (kubectl get configmap | grep -q "env-common"); then
+	kubectl delete configmap env-common
+fi
 kubectl create -f configMap-env-common.yaml
