+++ src/update-configMap-glance-setup.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-kubectl delete configmap glance-setup
+if (kubectl get configmap | grep -q "glance-setup"); then
+	kubectl delete configmap glance-setup
+fi
 kubectl create -f configMap-glance-setup.yaml
