+++ src/update-configMap-horizon-setup.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-kubectl delete configmap horizon-setup
+if (kubectl get configmap | grep -q "horizon-setup"); then
+	kubectl delete configmap horizon-setup
+fi
 kubectl create -f configMap-horizon-setup.yaml
