+++ src/update-configMap-keystone-setup.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-kubectl delete configmap keystone-setup
+if (kubectl get configmap | grep -q "keystone-setup"); then
+	kubectl delete configmap keystone-setup
+fi
 kubectl create -f configMap-keystone-setup.yaml
