+++ src/update-configMap-neutron-server-setup.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-kubectl delete configmap neutron-server-setup
+if (kubectl get configmap | grep -q "neutron-server-setup"); then
+	kubectl delete configmap neutron-server-setup
+fi
 kubectl create -f configMap-neutron-server-setup.yaml
